% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkPersonalIDBySSN.R
\name{getPrimaryPersonalID}
\alias{getPrimaryPersonalID}
\title{Takes Client, removes records where SSNs are invalid, then links the PersonalIDs
based upon matching SSNs.  The Client is returned with a column called PrimaryPersonalID
which should be a more unique identifier than the PersonalID alone.}
\usage{
getPrimaryPersonalID(client)
}
\arguments{
\item{dataframe}{path to folder contaning first CSV set.}
}
\description{
Takes Client, removes records where SSNs are invalid, then links the PersonalIDs
based upon matching SSNs.  The Client is returned with a column called PrimaryPersonalID
which should be a more unique identifier than the PersonalID alone.
}
\examples{

clientDf2 <- getPrimaryPersonalID(clientDf1)
# Remove the old PersonalID
clientDf2 <- within(clientDf2, rm(PersonalID))
colnames(clientDf2)[1] <- "PersonalID"
# Get a deduplicated Client
clientDf2 <- unique(clientDf2)
}
